backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
pyjwt==2.8.1
bcrypt==4.1.2
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
